- refactor: move tor subpackage under network subpackage
