+++ androidNode/src/androidMain/kotlin/network/bisq/mobile/android/node/service/network/tor/TorBridgeCommandHandler.kt
@@ -1,4 +1,4 @@
-package network.bisq.mobile.android.node.service.tor
+package network.bisq.mobile.android.node.service.network.tor
 
 import network.bisq.mobile.domain.utils.Logging
