+++ androidNode/src/androidTest/kotlin/network/bisq/mobile/android/node/service/tor/KmpTorControlPortTest.kt
@@ -3,6 +3,7 @@ package network.bisq.mobile.android.node.service.tor
 import android.content.Context
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
+import network.bisq.mobile.android.node.service.network.tor.TorService
 import org.junit.Assert
 import org.junit.Test
 import org.junit.runner.RunWith
